# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Laravel with PostgreSQL - Created Muhamad Ferdinal",
        "version": "1.0.0"
    },
    "paths": {
...
        "/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get list of users",
                "description": "Returns list of users",
                "operationId": "findAllUser",
                "responses": {
                    "200": {
                        "description": "Pengambilan Data User Berhasil!"
                    },
                    "500": {
                        "description": "Terjadi Kesalahan Saat Pengambilan Data User!"
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create user data",
                "description": "Create user data",
                "operationId": "createUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_depan",
                                    "nama_belakang",
                                    "jk",
                                    "tgl_lahir"
                                ],
                                "properties": {
                                    "nama_depan": {
                                        "description": "Nama Depan User (Required)",
                                        "type": "string"
                                    },
                                    "nama_belakang": {
                                        "description": "Nama Belakang User (Required)",
                                        "type": "string"
                                    },
                                    "alamat": {
                                        "description": "Alamat User (Nullable)",
                                        "type": "string"
                                    },
                                    "nomor_hp": {
                                        "description": "Nomor Handphone User (Nullable)",
                                        "type": "integer"
                                    },
                                    "jk": {
                                        "description": "Jenis Kelamin User (Required)",
                                        "type": "string"
                                    },
                                    "tgl_lahir": {
                                        "description": "Tanggal Lahir User (Required)",
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data User Berhasil Dibuat!"
                    },
                    "500": {
                        "description": "Terjadi Kesalahan Saat Membuat Data User!"
                    }
                }
            }
        },
        "/user/{user_id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get single user data",
                "description": "Returns single user data",
                "operationId": "findOneUser",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pengambilan Data User Berhasil!"
                    },
                    "404": {
                        "description": "Data User Tidak Ditemukan!"
                    },
                    "500": {
                        "description": "Terjadi Kesalahan Saat Pengambilan Data User!"
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update user data",
                "description": "Update user data",
                "operationId": "updateOneUser",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama_depan": {
                                        "description": "Nama Depan User (Optional)",
                                        "type": "string"
                                    },
                                    "nama_belakang": {
                                        "description": "Nama Belakang User (Optional)",
                                        "type": "string"
                                    },
                                    "alamat": {
                                        "description": "Alamat User (Optional)",
                                        "type": "string"
                                    },
                                    "nomor_hp": {
                                        "description": "Nomor Handphone User (Optional)",
                                        "type": "integer"
                                    },
                                    "jk": {
                                        "description": "Jenis Kelamin User (Optional)",
                                        "type": "string"
                                    },
                                    "tgl_lahir": {
                                        "description": "Tanggal Lahir User (Optional)",
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data User Berhasil Diperbarui!"
                    },
                    "404": {
                        "description": "Data User Tidak Ditemukan!"
                    },
                    "500": {
                        "description": "Terjadi Kesalahan Saat Memperbarui Data User!"
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete user data",
                "description": "Delete user data",
                "operationId": "deleteOneUser",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data User Berhasil Dihapus!"
                    },
                    "404": {
                        "description": "Data User Tidak Ditemukan!"
                    },
                    "500": {
                        "description": "Terjadi Kesalahan Saat Menghapus Data User!"
                    }
                }
            }
        }
    }
}